enable unicode password support
