+++ src/Makevars.in
@@ -1,9 +1,5 @@
 PKG_CPPFLAGS = @cflags@ -I. -Icryptography -Icryptography/msoc/include
-PKG_LIBS = -Lcryptography -lmsoc @libs@
-
-LIBMSOC = cryptography/msoc/msocdll.o
-
-STATLIB = cryptography/libmsoc.a
+PKG_LIBS = -Lcryptography @libs@
 
 # PKG_CXXFLAGS= -Wall -pedantic
 all: clean
@@ -11,8 +7,5 @@ all: clean
 all: $(SHLIB)
 $(SHLIB): $(STATLIB)
 
-$(STATLIB): $(LIBMSOC)
-	$(AR) rcs cryptography/libmsoc.a $(LIBMSOC)
-
 clean:
-	rm -f $(OBJECTS) $(SHLIB) $(STATLIB) $(LIBMSOC)
+	rm -f $(OBJECTS) $(SHLIB) $(STATLIB)
